+++ NEWS.md
@@ -1,4 +1,7 @@
 # OmopSketch 0.2.2.900
 
 -   Account for int64 in summariseInObservation by @cecicampanile #311
+
 -   eunomiaIsAvailable instead of the deprecated eunomia_is_available #316
+
+-   Only records in observation are accounted in summariseConceptIdCounts and summariseConceptSetCounts by @cecicampanile #319